Kept _ping() For Compatibility
Re-added _ping to prevent having to move to a major version. Some users may have used/relied on this pseudo-private method and I do not want to break anyones implementations. We could deprecate it but its not really worth the effort since `_ping` is just an alias to the public `ping` method.

Also updated tests and API docs to reflect the re-adding of _ping.